requirements.txt
absl-py==2.1.0
alembic==1.14.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.8.2
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
colorful==0.5.6
contourpy==1.3.0
cyaron==0.6.0
cycler==0.12.1
distro==1.9.0
docutils==0.21.2
easygui==0.98.3
et_xmlfile==2.0.0
fastapi==0.115.8
Flask==3.0.3
Flask-SocketIO==5.5.1
flatbuffers==25.1.24
fonttools==4.53.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httpx==0.27.2
id==1.5.0
idna==3.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jiter==0.5.0
keras==3.8.0
keyring==25.6.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.6.0
mss==9.0.2
namex==0.0.8
nh3==0.2.20
numpy==2.0.2
openai==1.45.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==6.1.1
pycryptodome==3.21.0
pydantic==2.9.1
pydantic_core==2.23.3
pygame==2.6.1
Pygments==2.19.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pynput==1.7.7
pyparsing==3.1.4
pyperclip==1.9.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.0
readme_renderer==44.0
redis==5.2.1
reflex==0.6.8
reflex-chakra==0.6.3
reflex-hosting-cli==0.1.34
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
setuptools==75.1.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.45.3
starlette-admin==0.14.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tomlkit==0.13.2
tqdm==4.66.5
ttkbootstrap==1.10.1
twine==6.1.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.0.4
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xeger==0.4.0
